+++ server/crawler.js
@@ -5,11 +5,14 @@
   const parsermaxredirects = 5;
   const mongotimeout = 5;
 
-  const log = false;
-
-  const os = require('os');
   const fs = require('fs');
+  const moment = require('moment');
+  const request = require('request');
+  const async = require('async');
+  const MongoClient = require('mongodb').MongoClient;
   const Parser = require('rss-parser');
+
+  const url = 'mongodb://127.0.0.1:27017/opentitles';
   const parser = new Parser({
     headers: {'User-Agent': 'OpenTitles Scraper by floris@debijl.xyz'},
     timeout: parsertimeout * 1000,
@@ -18,11 +21,6 @@
       item: ['wp:arc_uuid'],
     },
   });
-  const moment = require('moment');
-  const request = require('request');
-  const MongoClient = require('mongodb').MongoClient;
-  const url = 'mongodb://localhost:27017/opentitles';
-
   const listeners = [
     {
       name: 'NOSEdits',
@@ -32,9 +30,8 @@
   ];
 
   if (!fs.existsSync('media.json')) {
-    logger('FATAL - Media.json could not be found in the server directory.');
     throw new Error('Media.json could not be found in the server directory.');
-  } const CONFIG = JSON.parse(fs.readFileSync('media.json'));
+  } const config = JSON.parse(fs.readFileSync('media.json'));
 
   let dbo;
 
@@ -45,78 +42,97 @@
     appname: 'OpenTitles API',
     useNewUrlParser: true,
     connectTimeoutMS: mongotimeout * 1000,
-  }, function(err, database) {
-    if (err) {
-      throw err;
-    }
-
-    dbo = database.db('opentitles');
+  }).then((client) => {
+    dbo = client.db('opentitles');
+  }).catch((err) => {
+    throw err;
   });
-
-  // Do first retrieval of articles
-  setTimeout(() => {
-    logger('Starting OpenTitles Crawler...', 'Settings:', `Parser timeout: ${parsertimeout}`, `Parser max redirects: ${parsermaxredirects}`, `MongoDB Timeout: ${mongotimeout}`, `Scrape Interval: ${CONFIG.SCRAPER_INTERVAL}`);
-    retrieveArticles();
-  }, 5000);
-
-  setInterval(() => {
-    retrieveArticles();
-  }, CONFIG.SCRAPER_INTERVAL * 1000);
 
   /**
    * Iterate over all RSS feeds and check for each article if we've seen it already or not
    */
-  async function retrieveArticles() {
-    for (const countrykey in CONFIG.FEEDS) {
-      if (CONFIG.FEEDS.hasOwnProperty(countrykey)) {
-        const countryfeeds = CONFIG.FEEDS[countrykey];
-        for (let i = 0; i < countryfeeds.length; i++) {
-          const org = countryfeeds[i];
-          const orgfeed = {items: []};
-          for (let j = 0; j < org.FEEDS.length; j++) {
-            const feedname = org.FEEDS[j];
-            const [err, feed] = await to(parser.parseURL(org.PREFIX + feedname + org.SUFFIX));
+  function retrieveArticles() {
+    // console.log('Starting retrieve articles.');
+    Object.entries(config.feeds).forEach((country) => {
+      const countrycode = country[0];
+      const media = country[1];
 
-            if (err) {
-              logger(`Could not retrieve ${org.PREFIX + feedname + org.SUFFIX}`);
-              continue;
-            }
+      async.forEachSeries(media, (medium, nextMedium) => {
+        let mediumfeed = {items: []};
 
-            feed.items = feed.items.map((item) => {
-              item.artid = guidReducer(item[org.ID_CONTAINER], org.ID_MASK);
-
-              if (!item.artid) {
-                return false;
-              }
-
-              item.org = org.NAME;
-              item.feedtitle = feed.title;
-              item.sourcefeed = feedname;
-              item.lang = countrykey;
-              return item;
-            });
-
-            // Remove articles for which no guid exists or none was found
-            feed.items = feed.items.filter(Boolean);
-
-            orgfeed.items.push(...feed.items);
+        async.forEachSeries(medium.feeds, (feedname, nextFeed) => {
+          parser.parseURL(medium.prefix + feedname + medium.suffix)
+              .then((feed) => {
+                feed = processFeed(feed, medium, feedname, countrycode);
+                mediumfeed.items.push(...feed.items);
+                nextFeed();
+              })
+              .catch((err) => {
+                // console.log(`Could not retrieve ${medium.prefix + feedname + medium.suffix}`);
+                nextFeed();
+              });
+        }, (err) => {
+          // Callback function once all feeds are processed.
+          if (err) {
+            // Something went wrong when retrieving the feeds.
+            console.error(err);
+            return;
           }
 
-          removeDupesAndCheck(orgfeed);
+          mediumfeed = deduplicate(mediumfeed);
+          checkAndPropagate(mediumfeed);
+
+          nextMedium();
+        });
+      }, (err) => {
+        // Callback function once all media are processed.
+        if (err) {
+          // One the medium failed to process, do something here.
+          console.error(err);
         }
-      }
-    }
+      });
+    });
   }
 
   /**
-   * Takes a fully populated feed from one org, removes the duplicates and passes it on to the database to check for new titles.
-   * @param {object} feed The feed as retrieved by rss-parser - make sure artid and org are populated.
+   * Match a guid to each item for the subfeed and check for empty/invalid entries.
+   * @param {object} feed The subfeed as retrieved by rss-parser - make sure artid and org are populated.
+   * @param {object} medium The medium as defined in media.json
+   * @param {string} feedname The name of feed, which will be injected in the article.
+   * @param {string} countrycode The ISO 3166-1 Alpha-2 countrycode as defined in media.json
+   * @return {object} The feed with all extra variables injected and empty/invalid entries removed.
    */
-  function removeDupesAndCheck(feed) {
+  function processFeed(feed, medium, feedname, countrycode) {
+    feed.items = feed.items.map((item) => {
+      item.artid = guidReducer(item[medium.id_container], medium.id_mask);
+
+      if (!item.artid) {
+        return false;
+      }
+
+      item.org = medium.name;
+      item.feedtitle = feed.title;
+      item.sourcefeed = feedname;
+      item.lang = countrycode;
+      return item;
+    });
+
+    // Remove articles for which no guid exists or none was found
+    feed.items = feed.items.filter(Boolean);
+    return feed;
+  }
+
+  /**
+   * Takes a fully populated feed from one medium and removes the duplicates.
+   * @param {object} feed The orgfeed as retrieved by rss-parser - make sure artid and org are populated.
+   * @return {object} The mediumfeed without any duplicate entries.
+   */
+  function deduplicate(feed) {
     // Reduce feed items to unique ID's only
     const seen = {};
+
     feed.items = feed.items.filter((item) => {
-      // Required variables
+      // Make sure required variables are present
       if (!item.artid || !item.org) {
         return false;
       }
@@ -124,9 +140,20 @@
       return seen.hasOwnProperty(item.artid) ? false : (seen[item.artid] = true);
     });
 
+    return feed;
+  }
+
+  /**
+   * Send every item to be checked by the DB.
+   * @param {object} feed The orgfeed as retrieved by rss-parser - make sure artid and org are populated.
+   * @return {object} The same feed as was passed to checkAndPropagate().
+   */
+  function checkAndPropagate(feed) {
     feed.items.forEach((item) => {
       checkWithDB(item);
     });
+
+    return feed;
   }
 
   /**
@@ -141,6 +168,7 @@
 
     findArticle(find, (res) => {
       if (res === [null]) {
+        // console.log(`[${article.org}:${article.artid}] Not found in db`);
         return;
       }
 
@@ -188,7 +216,7 @@
   function findArticle(find, callback) {
     dbo.collection('articles').findOne(find, function(err, res) {
       if (err) {
-        logger(err);
+        // console.log(`[${moment().format('DD/MM/Y - HH:mm:ss')}] ${err}`);
         if (typeof(callback) == 'function') {
           callback([null]);
         }
@@ -197,12 +225,13 @@
 
       if (typeof(callback) == 'function') {
         callback(res);
+        return res;
       }
     });
   }
 
   /**
-   * Reduce a given GUID (effectively a URL) to a full ID used for tracking the article.
+   * Reduce a given GUID (usually a URL) to a full ID used for tracking the article.
    * @param {string} guid The GUID for this article.
    * @param {string} mask The regex mask to extract the ID with.
    * @return {string} The article ID contained within the GUID.
@@ -221,22 +250,18 @@
   }
 
   /**
-   * Await a promise and return its error if one occurs.
-   * @param {Promise} promise
-   * @return {[String, Promise]} An error (null if none occurred) and the result of the promise.
-   */
-  function to(promise) {
-    return promise.then((data) => {
-      return [null, data];
-    })
-        .catch((err) => [err]);
-  }
-
-  /**
    * Notify all defined listeners that the title for an article has changed.
    * @param {Object} article The article object.
    */
   function notifyListeners(article) {
+    if (!article.org || !article.articleID) {
+      return;
+    }
+
+    if (listeners.length === 0) {
+      return;
+    }
+
     listeners.forEach((listener) => {
       if (listener.interestedOrgs.includes(article.org)) {
         request.post({
@@ -245,40 +270,18 @@
           body: article,
         }, function(err, httpResponse, body) {
           if (err) {
-            logger(`Could not reach ${listener.name} when issuing webhook.`);
+            console.log(`[${moment().format('DD/MM/Y - HH:mm:ss')}] Could not reach ${listener.name} when issuing webhook.`);
+          } else {
+            console.log(`[${moment().format('DD/MM/Y - HH:mm:ss')}] Reached ${listener.name} for [${article.org}:${article.articleID}].`);
           }
         });
       }
     });
+
+    return;
   }
 
-  /**
-   * Write N number of lines to the logfiles preceded by a timestamp.
-   * @param  {...string} params The lines to write to the logfile
-   */
-  function logger(...params) {
-    if (!log) return;
-
-    const dir = '/var/log/opentitles';
-    const file = `${dir}/crawler.log`;
-    if (!fs.existsSync(dir)) {
-      fs.mkdirSync(dir);
-    }
-
-    let output = `[${moment().format('DD/MM/Y - HH:mm:ss')}] `;
-    const leftpadlength = [...output].length;
-
-    for (let i = 0; i < params.length; i++) {
-      if (i > 0) {
-        output += ' '.repeat(leftpadlength);
-      }
-
-      output += `${params[i]}${os.EOL}`;
-    }
-
-    fs.appendFile(file, output, (err) => {
-      if (!err) return;
-      console.error('Could not write to logfile! - ' + err);
-    });
-  }
+  setInterval(() => {
+    retrieveArticles();
+  }, config.scraper_interval * 1000);
 })();