+++ changelog.md
@@ -1,3 +1,12 @@
+### 1.9
+OpenTitles will now lookup the title history for an article via a background script proxy, instead of directly from the content script.
+This is more secure and is not blocked by future CORB-policies, such as those already in place in Chrome 73
+
+- Switched out XMLHttpRequest API usage for the newer, more robust fetch API.
+- Moved CORS request to background script.
+- Added floris.amsterdam (HTTPS only) as allowed host.
+- Properties in media.json are now lowercased as per best JSON-practices.
+
 ### 1.8
 NOS title selector changed